fix: Removed some extra code [#123]
