+++ src/modules/identity/services/team-user.service.ts
@@ -17,6 +17,7 @@ import * as nodemailer from "nodemailer";
 import hbs = require("nodemailer-express-handlebars");
 import { ConfigService } from "@nestjs/config";
 import path = require("path");
+import { EmailService } from "@src/modules/common/services/email.service";
 /**
  * Team User Service
  */
@@ -29,6 +30,7 @@ export class TeamUserService {
     private readonly producerService: ProducerService,
     private readonly teamService: TeamService,
     private readonly configService: ConfigService,
+    private readonly emailService: EmailService,
   ) {}
 
   async HasPermissionToRemove(
@@ -627,32 +629,57 @@ export class TeamUserService {
    * @param {string} email - The email address to send the notification to.
    * @returns {Promise<void>}
    */
+  // async leaveTeamEmail(
+  //   MemberName: string,
+  //   teamName: string,
+  //   OwnerName: string,
+  //   email: string,
+  // ): Promise<void> {
+  //   const transporter = nodemailer.createTransport({
+  //     host: this.configService.get("app.mailHost"),
+  //     port: this.configService.get("app.mailPort"),
+  //     secure: this.configService.get("app.mailSecure") === "true",
+  //     auth: {
+  //       user: this.configService.get("app.userName"),
+  //       pass: this.configService.get("app.senderPassword"),
+  //     },
+  //   });
+  //   const handlebarOptions = {
+  //     viewEngine: {
+  //       extname: ".handlebars",
+  //       partialsDir: path.resolve(__dirname, "..", "..", "views", "partials"),
+  //       layoutsDir: path.resolve(__dirname, "..", "..", "views", "layouts"),
+  //       defaultLayout: "main", // Use the main.handlebars layout
+  //     },
+  //     viewPath: path.resolve(__dirname, "..", "..", "views"),
+  //     extName: ".handlebars",
+  //   };
+  //   transporter.use("compile", hbs(handlebarOptions));
+  //   const mailOptions = {
+  //     from: this.configService.get("app.senderEmail"),
+  //     to: email,
+  //     text: "Leaving Team",
+  //     template: "leaveTeamEmail",
+  //     context: {
+  //       ownerName: OwnerName,
+  //       memberName: MemberName,
+  //       teamName: teamName,
+  //       sparrowEmail: this.configService.get("support.sparrowEmail"),
+  //     },
+  //     subject: `Team Member Update: ${MemberName} has left ${teamName}`,
+  //   };
+  //   const promise = [transporter.sendMail(mailOptions)];
+  //   await Promise.all(promise);
+  // }
+
   async leaveTeamEmail(
     MemberName: string,
     teamName: string,
     OwnerName: string,
     email: string,
   ): Promise<void> {
-    const transporter = nodemailer.createTransport({
-      host: this.configService.get("app.mailHost"),
-      port: this.configService.get("app.mailPort"),
-      secure: this.configService.get("app.mailSecure") === "true",
-      auth: {
-        user: this.configService.get("app.userName"),
-        pass: this.configService.get("app.senderPassword"),
-      },
-    });
-    const handlebarOptions = {
-      viewEngine: {
-        extname: ".handlebars",
-        partialsDir: path.resolve(__dirname, "..", "..", "views", "partials"),
-        layoutsDir: path.resolve(__dirname, "..", "..", "views", "layouts"),
-        defaultLayout: "main", // Use the main.handlebars layout
-      },
-      viewPath: path.resolve(__dirname, "..", "..", "views"),
-      extName: ".handlebars",
-    };
-    transporter.use("compile", hbs(handlebarOptions));
+    const transporter = this.emailService.createTransporter();
+
     const mailOptions = {
       from: this.configService.get("app.senderEmail"),
       to: email,
@@ -666,8 +693,8 @@ export class TeamUserService {
       },
       subject: `Team Member Update: ${MemberName} has left ${teamName}`,
     };
-    const promise = [transporter.sendMail(mailOptions)];
-    await Promise.all(promise);
+
+    await transporter.sendMail(mailOptions);
   }
 
   /**
@@ -684,26 +711,8 @@ export class TeamUserService {
     OwnerName: string,
     email: string,
   ): Promise<void> {
-    const transporter = nodemailer.createTransport({
-      host: this.configService.get("app.mailHost"),
-      port: this.configService.get("app.mailPort"),
-      secure: this.configService.get("app.mailSecure") === "true",
-      auth: {
-        user: this.configService.get("app.userName"),
-        pass: this.configService.get("app.senderPassword"),
-      },
-    });
-    const handlebarOptions = {
-      viewEngine: {
-        extname: ".handlebars",
-        partialsDir: path.resolve(__dirname, "..", "..", "views", "partials"),
-        layoutsDir: path.resolve(__dirname, "..", "..", "views", "layouts"),
-        defaultLayout: "main", // Use the main.handlebars layout
-      },
-      viewPath: path.resolve(__dirname, "..", "..", "views"),
-      extName: ".handlebars",
-    };
-    transporter.use("compile", hbs(handlebarOptions));
+    const transporter = this.emailService.createTransporter();
+
     const mailOptions = {
       from: this.configService.get("app.senderEmail"),
       to: email,
@@ -733,26 +742,8 @@ export class TeamUserService {
     OwnerName: string,
     email: string,
   ): Promise<void> {
-    const transporter = nodemailer.createTransport({
-      host: this.configService.get("app.mailHost"),
-      port: this.configService.get("app.mailPort"),
-      secure: this.configService.get("app.mailSecure") === "true",
-      auth: {
-        user: this.configService.get("app.userName"),
-        pass: this.configService.get("app.senderPassword"),
-      },
-    });
-    const handlebarOptions = {
-      viewEngine: {
-        extname: ".handlebars",
-        partialsDir: path.resolve(__dirname, "..", "..", "views", "partials"),
-        layoutsDir: path.resolve(__dirname, "..", "..", "views", "layouts"),
-        defaultLayout: "main", // Use the main.handlebars layout
-      },
-      viewPath: path.resolve(__dirname, "..", "..", "views"),
-      extName: ".handlebars",
-    };
-    transporter.use("compile", hbs(handlebarOptions));
+    const transporter = this.emailService.createTransporter();
+
     const mailOptions = {
       from: this.configService.get("app.senderEmail"),
       to: email,
@@ -781,26 +772,8 @@ export class TeamUserService {
     OwnerName: string,
     email: string,
   ): Promise<void> {
-    const transporter = nodemailer.createTransport({
-      host: this.configService.get("app.mailHost"),
-      port: this.configService.get("app.mailPort"),
-      secure: this.configService.get("app.mailSecure") === "true",
-      auth: {
-        user: this.configService.get("app.userName"),
-        pass: this.configService.get("app.senderPassword"),
-      },
-    });
-    const handlebarOptions = {
-      viewEngine: {
-        extname: ".handlebars",
-        partialsDir: path.resolve(__dirname, "..", "..", "views", "partials"),
-        layoutsDir: path.resolve(__dirname, "..", "..", "views", "layouts"),
-        defaultLayout: "main", // Use the main.handlebars layout
-      },
-      viewPath: path.resolve(__dirname, "..", "..", "views"),
-      extName: ".handlebars",
-    };
-    transporter.use("compile", hbs(handlebarOptions));
+    const transporter = this.emailService.createTransporter();
+
     const mailOptions = {
       from: this.configService.get("app.senderEmail"),
       to: email,
@@ -832,28 +805,7 @@ export class TeamUserService {
     userName: string,
     email: string,
   ): Promise<void> {
-    const transporter = nodemailer.createTransport({
-      host: this.configService.get("app.mailHost"),
-      port: this.configService.get("app.mailPort"),
-      secure: this.configService.get("app.mailSecure") === "true",
-      auth: {
-        user: this.configService.get("app.userName"),
-        pass: this.configService.get("app.senderPassword"),
-      },
-    });
-
-    const handlebarOptions = {
-      viewEngine: {
-        extname: ".handlebars",
-        partialsDir: path.resolve(__dirname, "..", "..", "views", "partials"),
-        layoutsDir: path.resolve(__dirname, "..", "..", "views", "layouts"),
-        defaultLayout: "main", // Use the main.handlebars layout
-      },
-      viewPath: path.resolve(__dirname, "..", "..", "views"),
-      extName: ".handlebars",
-    };
-
-    transporter.use("compile", hbs(handlebarOptions));
+    const transporter = this.emailService.createTransporter();
 
     const mailOptions = {
       from: this.configService.get("app.senderEmail"),
@@ -887,28 +839,7 @@ export class TeamUserService {
     userName: string,
     email: string,
   ): Promise<void> {
-    const transporter = nodemailer.createTransport({
-      host: this.configService.get("app.mailHost"),
-      port: this.configService.get("app.mailPort"),
-      secure: this.configService.get("app.mailSecure") === "true",
-      auth: {
-        user: this.configService.get("app.userName"),
-        pass: this.configService.get("app.senderPassword"),
-      },
-    });
-
-    const handlebarOptions = {
-      viewEngine: {
-        extname: ".handlebars",
-        partialsDir: path.resolve(__dirname, "..", "..", "views", "partials"),
-        layoutsDir: path.resolve(__dirname, "..", "..", "views", "layouts"),
-        defaultLayout: "main", // Use the main.handlebars layout
-      },
-      viewPath: path.resolve(__dirname, "..", "..", "views"),
-      extName: ".handlebars",
-    };
-
-    transporter.use("compile", hbs(handlebarOptions));
+    const transporter = this.emailService.createTransporter();
 
     const mailOptions = {
       from: this.configService.get("app.senderEmail"),
