+++ src/modules/views/inviteWorkspaceEmail.handlebars
@@ -1,127 +1,127 @@
-  {{!< layoutName}} <!DOCTYPE html>
-    <html lang="en">
+{{!< layoutName}} <!DOCTYPE html>
+  <html lang="en">
 
-    <head>
-      <meta charset="UTF-8">
-      <meta name="viewport" content="width=device-width, initial-scale=1.0">
-      <title>Invite Team</title>
-
-
-      <style>
-        body {
-          font-family: "Inter", sans-serif;
-          box-sizing: border-box;
-          padding: 0;
-          margin: 0;
-          transition: background-color 0.3s, color 0.3s;
-        }
-
-        .container {
-          display: flex;
-          justify-content: center;
-          width: 100%;
-          height: 100%;
-
-        }
-
-        .container-data {
-          margin: auto;
-          max-width: 570px;
-          background-color: var(--background-color, white);
-          padding: 20px;
-          transition: background-color 0.3s, color 0.3s;
-        }
-
-        .header-container {
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          gap: 10px;
-        }
-
-        .sparrow-heading-text {
-          font-size: 40px;
-          font-weight: 800;
-          color: var(--header-text-color, #3670F7);
-        }
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Invite Team</title>
 
 
+    <style>
+      body {
+        font-family: "Inter", sans-serif;
+        box-sizing: border-box;
+        padding: 0;
+        margin: 0;
+        transition: background-color 0.3s, color 0.3s;
+      }
 
-        .email-text-style {
-          color: var(--email-text-color, #333);
-          font-family: "Inter", sans-serif;
-          font-size: 14px;
-          font-style: normal;
-          font-weight: 400;
-          line-height: 21px;
-        }
+      .container {
+        display: flex;
+        justify-content: center;
+        width: 100%;
+        height: 100%;
 
-        .text-bold-dynamic {
-          color: var(--text-bold-color, #333333);
-          font-weight: 700;
-          line-height: 21px;
-        }
+      }
 
-        .icon {
-          transition: filter 0.3s;
-        }
+      .container-data {
+        margin: auto;
+        max-width: 570px;
+        background-color: var(--background-color, white);
+        padding: 20px;
+        transition: background-color 0.3s, color 0.3s;
+      }
 
-        .dark-mode .icon {
-          filter: invert(100%);
-        }
-      </style>
-    </head>
+      .header-container {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        gap: 10px;
+      }
 
-    <body>
-
-      <div class="container">
-        <section style="width: 100%;">
-          <div class="container-data">
-
-            {{> header}}
-
-            <section>
-
-              <div style="margin-bottom: 5%">
-                <h2 class="email-text-style">
-                  Hi {{firstname}},
-                </h2>
-
-                <p class="welcome-text email-text-style">
-                  Congratulations! {{firstname}} has invited you to join the {{workspacename}} workspace on Sparrow as
-                  {{userRole}}.
-                </p>
-
-                <p class="welcome-text email-text-style">
-                  Your creativity and input will help shape the success in this workspace. Log in to Sparrow now to
-                  start collaborating and making an impact to the team and the workspace!
-                </p>
-
-                <p>If you need additional assistance or you did not make this change, please contact us at <a
-                    href="mailto:{{sparrowEmail}}" style="text-decoration-line: underline;">{{sparrowEmail}}</a>
-                </p>
-
-                <p class="welcome-text email-text-style">
-                  Warm regards, <br> The Sparrow Team <br>
-                  <a href="https://sparrowapp.dev"
-                    style="font-weight: 400; text-decoration: none; color: #3670F7; border-bottom: 1px solid #3670F7 ;">
-                    https://sparrowapp.dev</a>
-                </p>
-
-              </div>
+      .sparrow-heading-text {
+        font-size: 40px;
+        font-weight: 800;
+        color: var(--header-text-color, #3670F7);
+      }
 
 
-              <hr>
 
-            </section>
+      .email-text-style {
+        color: var(--email-text-color, #333);
+        font-family: "Inter", sans-serif;
+        font-size: 14px;
+        font-style: normal;
+        font-weight: 400;
+        line-height: 21px;
+      }
+
+      .text-bold-dynamic {
+        color: var(--text-bold-color, #333333);
+        font-weight: 700;
+        line-height: 21px;
+      }
+
+      .icon {
+        transition: filter 0.3s;
+      }
+
+      .dark-mode .icon {
+        filter: invert(100%);
+      }
+    </style>
+  </head>
+
+  <body>
+
+    <div class="container">
+      <section style="width: 100%;">
+        <div class="container-data">
+
+          {{> header}}
+
+          <section>
+
+            <div style="margin-bottom: 5%">
+              <h2 class="email-text-style">
+                Hi {{firstname}},
+              </h2>
+
+              <p class="welcome-text email-text-style">
+                Congratulations! {{firstname}} has invited you to join the {{workspacename}} workspace on Sparrow as
+                {{!-- {{userRole}}. --}}
+              </p>
+
+              <p class="welcome-text email-text-style">
+                Your creativity and input will help shape the success in this workspace. Log in to Sparrow now to
+                start collaborating and making an impact to the team and the workspace!
+              </p>
+
+              <p>If you need additional assistance or you did not make this change, please contact us at <a
+                  href="mailto:{{sparrowEmail}}" style="text-decoration-line: underline;">{{sparrowEmail}}</a>
+              </p>
+
+              <p class="welcome-text email-text-style">
+                Warm regards, <br> The Sparrow Team <br>
+                <a href="https://sparrowapp.dev"
+                  style="font-weight: 400; text-decoration: none; color: #3670F7; border-bottom: 1px solid #3670F7 ;">
+                  https://sparrowapp.dev</a>
+              </p>
+
+            </div>
 
 
-            {{> footer}}
+            <hr>
+
+          </section>
 
 
-          </div>
-        </section>
-      </div>
-    </body>
+          {{> footer}}
 
-    </html>
+
+        </div>
+      </section>
+    </div>
+  </body>
+
+  </html>
